pyproject.toml
[project]
name = "tiny-git-cli-tools"
version = "0.1.0"
description = "Simple command-line tools for working with Git"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "gitpython>=3.1.0",
    "toml>=0.10.0",
    "openai>=0.27.0",
    "pydantic>=1.8.0",
    "pygithub>=2.8.1",
]

[project.scripts]
git-normalize-newlines = "tiny_git_cli_tools.git_normalize_newlines:main"
git-change-author = "tiny_git_cli_tools.git_change_author:main"
git-add-missing-signatures = "tiny_git_cli_tools.git_add_missing_signatures:main"
git-remote-change-protocol = "tiny_git_cli_tools.git_remote_change_protocol:main"
git-name-feature-branch = "tiny_git_cli_tools.git_name_feature_branch:main"
github-create-pr = "tiny_git_cli_tools.github_create_pr:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["tiny_git_cli_tools*"]
